backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==43.0.1
defusedxml==0.8.0rc2
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==4.2.16
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-configurations==2.5.1
django-cors-headers==4.4.0
django-filter==24.3
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==29.0.0
google-auth==2.35.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-factoryboy==2.7.0
pytest-mock==3.14.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
